requirements.txt
alembic==1.2.1
astroid==2.2.5
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cssselect2==0.2.2
cycler==0.10.0
Cython==0.29.13
idna==2.8
isort==4.3.21
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
lxml==4.4.1
Mako==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
numpy==1.17.4
pandas==0.25.3
Pillow==6.2.1
pygantt==0.1.2
pylint==2.3.1
pyparsing==2.4.5
pyTelegramBotAPI==3.6.6
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.3
reportlab==3.5.32
requests==2.22.0
schedule==0.6.0
six==1.12.0
SQLAlchemy==1.3.6
svglib==0.9.3
tinycss2==1.0.2
typed-ast==1.4.0
urllib3==1.25.3
vedis==0.7.1
webencodings==0.5.1
wrapt==1.11.2
